move add_plugin_to_db to database folder
